access control and delete permissions

+++ src/controllers/invParticipationController.js
@@ -1,7 +1,8 @@
 const controller = {};
 
 controller.list = (req, res) => {
-  let info = {};
+  if (req.session.loggedin) {
+    let info = {};
   req.getConnection((err, conn) => {
     conn.query('SELECT * FROM invsoc_participation', (err, invParticipations) => {
      if (err) {
@@ -48,8 +49,11 @@ controller.list = (req, res) => {
            }
         });
         
-      });
-}
+      })
+  } else {
+    res.redirect("/");
+  }
+}; 
 
 controller.save = (req, res) => {
   const data = req.body;
@@ -61,7 +65,10 @@ controller.save = (req, res) => {
   if(typeof data.application_state != "string"){
   data.application_state = `|${data.application_state.join(",")}|`;
   }
-  
+
+  if(typeof data.fort_theme != "string"){
+  data.fort_theme = `|${data.fort_theme.join(",")}|`;  
+  }
 
   console.log(req.body);
   req.getConnection((err, conn) => {
@@ -119,6 +126,10 @@ controller.update = (req, res) => {
   newParticipation.application_state = `|${newParticipation.application_state.join(",")}|`;
   }
 
+  if(typeof newParticipation.fort_theme != "string"){
+  newParticipation.fort_theme = `|${newParticipation.fort_theme.join(",")}|`;
+  }
+
   req.getConnection((err, conn) => {
 
   conn.query('UPDATE invsoc_participation set ? where id_partIS = ?', [newParticipation, id_partIS], (err, rows) => {
@@ -136,13 +147,21 @@ controller.cancel = (res) => {
 }
 
 controller.delete = (req, res) => {
-  const { id_partIS } = req.params;
-  req.getConnection((err, conn) => {
-    conn.query('DELETE FROM invsoc_participation WHERE id_partIS = ?', [id_partIS] , (err, rows) => {
-      console.log(rows);
-      res.redirect('/invParticipation');
-    });
-  });
+  var email = req.session.email;
+
+  if(email == 'v.cardin@fundacionmerded.org.mx'){
+    const { id_partIS } = req.params;
+    req.getConnection((err, conn) => {
+      conn.query('DELETE FROM invsoc_participation WHERE id_partIS = ?', [id_partIS] , (err, rows) => {
+        console.log(rows);
+        res.redirect('/invParticipation');
+      });
+    })
+  }else {
+    res.send("<script>alert('No cuenta con permiso para eliminar este registro'); window.location.href = '/invParticipation'; </script>");
+  }
+
+ 
 }
 
 module.exports = controller;
